+++ src/primitives/progress/index.ts
@@ -1,12 +1,16 @@
 export {
   ProgressContextProvider,
+  ProgressContextView,
   useProgressContext,
 } from "./context.js";
 export type {
   ProgressContextValue,
+  ProgressContextProps,
 } from "./context.js";
-export { ProgressRoot } from "./ProgressRoot.js";
-export type { ProgressRootProps } from "./ProgressRoot.js";
+export { ProgressRoot, ProgressRootProvider } from "./ProgressRoot.js";
+export type { ProgressRootProps, ProgressRootProviderProps } from "./ProgressRoot.js";
+export { useProgress } from "./useProgress.js";
+export type { UseProgressProps, ProgressController, ProgressIds } from "./useProgress.js";
 export { ProgressIndicator } from "./ProgressIndicator.js";
 export type { ProgressIndicatorProps } from "./ProgressIndicator.js";
 export {